requirements.txt
Django==1.5.5
simplejson==3.16.0
cassandra-driver==3.15.1
loremipsum==1.0.5
